requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
langchain
langchain-openai
requests
